# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jikan API",
...
        "termsOfService": "https://jikan.moe/terms",
        "contact": {
            "name": "API Support (Discord)",
            "url": "http://discord.jikan.moe"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jikan-me/jikan-rest/blob/master/LICENSE"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://api.jikan.moe/v4",
            "description": "Jikan REST API Beta"
        }
    ],
    "paths": {
...
        "/clubs/{id}": {
            "get": {
                "tags": [
                    "clubs"
                ],
                "operationId": "getClubsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Club Resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/club"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/clubs/{id}/members": {
            "get": {
                "tags": [
                    "clubs"
                ],
                "operationId": "getClubMembers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Club Members Resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/pagination"
                                        },
                                        {
                                            "$ref": "#/components/schemas/club_member"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/clubs/{id}/staff": {
            "get": {
                "tags": [
                    "clubs"
                ],
                "operationId": "getClubStaff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Club Staff",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/club_staff"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/clubs/{id}/relations": {
            "get": {
                "tags": [
                    "clubs"
                ],
                "operationId": "getClubRelations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Club Relations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/club_relations"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "club_member": {
                "description": "Club Member",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/user_images"
                        }
                    }
                },
                "type": "object"
            },
...
            "club_relations": {
                "description": "Club Relations",
                "properties": {
                    "data": {
                        "properties": {
                            "anime": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/mal_url"
                                }
                            },
                            "manga": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/mal_url"
                                }
                            },
                            "characters": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/mal_url"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "club": {
                "description": "Club Resource",
                "properties": {
                    "data": {
                        "properties": {
                            "mal_id": {
                                "description": "MyAnimeList ID",
                                "type": "integer"
                            },
                            "name": {
                                "description": "Club name",
                                "type": "string"
                            },
                            "url": {
                                "description": "Club URL",
                                "type": "string"
                            },
                            "images": {
                                "$ref": "#/components/schemas/common_images"
                            },
                            "members": {
                                "description": "Number of club members",
                                "type": "integer"
                            },
                            "category": {
                                "description": "Club Category",
                                "type": "string",
                                "enum": [
                                    "actors & artists",
                                    "anime",
                                    "characters",
                                    "cities & neighborhoods",
                                    "companies",
                                    "conventions",
                                    "games",
                                    "japan",
                                    "manga",
                                    "music",
                                    "others",
                                    "schools"
                                ]
                            },
                            "created": {
                                "description": "Date Created ISO8601",
                                "type": "string"
                            },
                            "access": {
                                "description": "Club access",
                                "type": "string",
                                "enum": [
                                    "public",
                                    "private",
                                    "secret"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "club_staff": {
                "description": "Club Staff Resource",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "description": "User URL",
                                    "type": "string"
                                },
                                "username": {
                                    "description": "User's username",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "mal_url": {
                "description": "Parsed URL Data",
                "properties": {
                    "mal_id": {
                        "description": "MyAnimeList ID",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Type of resource",
                        "type": "string"
                    },
                    "name": {
                        "description": "Resource Name/Title",
                        "type": "string"
                    },
                    "url": {
                        "description": "MyAnimeList URL",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "pagination": {
                "properties": {
                    "pagination": {
                        "properties": {
                            "last_visible_page": {
                                "type": "integer"
                            },
                            "has_next_page": {
                                "type": "boolean"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "user_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "webp": {
                        "description": "Available images in WEBP",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL WEBP",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "common_images": {
                "properties": {
                    "jpg": {
                        "description": "Available images in JPG",
                        "properties": {
                            "image_url": {
                                "nullable": true,
                                "description": "Image URL JPG",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
...
        }
    },
...
}